fix: search existing, markdown rendering
