+++ .changeset/did-you-know-that-the-world-is-round.md
@@ -1,0 +1,5 @@
+---
+'@node-core/doc-kit': patch
+---
+
+Close Orama search when the target link is on the same page
