+++ .changeset/yes-i-did-know-that-thank-you.md
@@ -1,0 +1,5 @@
+---
+'@node-core/doc-kit': patch
+---
+
+Render markdown `code` snippets in the sidebar
